+++ Data/Base.rte/Activities/OneManArmyZeroG.lua
@@ -86,6 +86,8 @@ function OneManArmyZeroG:StartNewGame()
 	self.timeDisplay = timeLimitText .. " minutes";
 	self.enemySpawnTimeLimit = 500;
 	
+	MusicMan:PlayDynamicSong("Generic Battle Music");
+	
 	local automoverController = CreateActor("Invisible Automover Controller", "Base.rte");
 	automoverController.Team = -1;
 	automoverController:SetNumberValue("MovementSpeed", 16);
@@ -268,16 +270,12 @@ function OneManArmyZeroG:EndActivity()
 	if not self:IsPaused() then
 		--Play sad music if no humans are left
 		if self:HumanBrainCount() == 0 then
-			AudioMan:ClearMusicQueue();
-			AudioMan:PlayMusic("Base.rte/Music/dBSoundworks/udiedfinal.ogg", 2, -1.0);
-			AudioMan:QueueSilence(10);
-			AudioMan:QueueMusicStream("Base.rte/Music/dBSoundworks/ccambient4.ogg");
+			MusicMan:PlayDynamicSong("Generic Defeat Music", "Default", true);
+			MusicMan:PlayDynamicSong("Generic Ambient Music");
 		else
 			--But if humans are left, play happy music!
-			AudioMan:ClearMusicQueue();
-			AudioMan:PlayMusic("Base.rte/Music/dBSoundworks/uwinfinal.ogg", 2, -1.0);
-			AudioMan:QueueSilence(10);
-			AudioMan:QueueMusicStream("Base.rte/Music/dBSoundworks/ccambient4.ogg");
+			MusicMan:PlayDynamicSong("Generic Victory Music", "Default", true);
+			MusicMan:PlayDynamicSong("Generic Ambient Music");
 		end
 	end
 end